+++ src/main/groovy/com/exictos/devops/helpers/LiberoHelper.groovy
@@ -19,7 +19,7 @@ class LiberoHelper {
 
     protected XHDLogger log = new Application().getLog()
 
-    private static final String DATE_FORMAT = "yyyy-MM-dd_HH-mm-ss"
+    static final String DATE_FORMAT = "yyyy-MM-dd_HH-mm-ss"
     private static final String NAME_VALIDATOR_REGEX = "^([aA-zZ0-9\\-]+)((_v)(?:[\\dx]{1,3}\\.){0,3}[\\dx]{1,3})?(___)"
     private static final String VERSION_VALIDATOR_REGEX = "^(?:[\\dx]{1,3}\\.){0,3}[\\dx]{1,3}\$"
 
@@ -80,7 +80,7 @@ class LiberoHelper {
             throw new IllegalArgumentException("Application name provided ${standardizedName} does not contain a version number.")
 
         try{
-            String version = standardizedName.substring(standardizedName.indexOf("_v"), standardizedName.indexOf("___"))
+            String version = standardizedName.substring(standardizedName.indexOf("___v") + 4, standardizedName.lastIndexOf("___"))
             return version
         }catch(Exception e){
             log.log("Could not parse name ${standardizedName}. Cause: ${e}")